pyproject.toml
[project]
name = "sap-data-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "flask-session>=0.8.0",
    "python-dotenv>=1.0.1",
]
